move `ize tunnel` functionality to `ize tunnel up` move all functionality of `ize tunnel` to `ize tunnel up`. Also remove `ize tunnel ssh-key` command.
